Do not mark placeholder text as safe
Closes #134

Placeholder values are run through `mark_safe` when building form
widgets. That's not a good idea, as it permits XSS content. As
placeholders are often generated automatically from labels, and labels
are properly escaped, this is unexpected behaviour.

This behaviour was introduced in
5b98fd5e5a750bb8588eb92569db324b736c736e. The commit doesn't give a
reason for the introduction (and neither does the PR at
https://github.com/dyve/django-bootstrap3/pull/385).