requirements: Accept importlib-metadata 2.x.x
A new version of importlib-metadata (2.0.0) has been released a few
days ago. Its changelog [1] mentions that the only change is the
removal of `importlib_metadata.__version__`, which is not used by
django-bootstrap4. So we can safely accept this version.

The current pinning (^1.5.0) is problematic because projects that
depend on django-bootstrap4 may require importlib-metadata with a more
laxist pinning (or no pinning at all).

[1] https://importlib-metadata.readthedocs.io/en/latest/changelog%20%28links%29.html#v2-0-0